refactor: PR #40 리뷰 반영 — 상수 분리, 더블클릭 방어, z-index 문서 수정, 스타일 호이스팅, useFadeIn 공통화
- 로그아웃 상수(LOGOUT_DELAY_MS, LOGOUT_FADE_DURATION_MS)를 constants/logout.ts로 분리
- handleLogout에 isLoggingOut 가드 추가하여 더블 클릭 방어
- 설계 문서 z-index 값을 구현과 일치하도록 99999로 수정
- WindowRouter 인라인 스타일을 모듈 레벨 상수로 호이스팅
- DesktopPage/LoginPage의 fade-in 패턴을 useFadeIn 커스텀 훅으로 공통화
