+++ src/pages/DesktopPage/constants/logout.ts
@@ -1,0 +1,5 @@
+/** 오버레이 체류 시간 (ms) */
+export const LOGOUT_DELAY_MS = 1000;
+
+/** 페이드인 애니메이션 시간 (ms) */
+export const LOGOUT_FADE_DURATION_MS = 400;
