+++ src/shared/hooks/index.ts
@@ -1,4 +1,5 @@
 export { default as useAxios } from "./useAxios";
+export { default as useFadeIn } from "./useFadeIn";
 export { default as useGetCurrentTime } from "./useGetCurrentTime";
 export { default as useOutsideClick } from "./useOutsideClick";
 export { default as useScreenHeight } from "./useScreenHeight";
